[C-1559] Fix audio-breakdown-drawer
