+++ packages/mobile/src/components/collectible-details-drawer/CollectibleDetailsDrawer.tsx
@@ -104,6 +104,12 @@ export const CollectibleDetailsDrawer = () => {
 
   const handleShare = useCallback(() => {
     if (owner && collectible) {
+      console.log(
+        'handle share',
+        owner.handle,
+        collectible.id,
+        getCollectiblesRoute(owner.handle, collectible.id)
+      )
       const url = getCollectiblesRoute(owner.handle, collectible.id)
       share({ url })
     }
